+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `fep-e232` feature flag (disabled by default).
 
+### Fixed
+
+- Make webclient-to-object redirects work for remote profiles and posts.
+
 ## [1.17.0] - 2023-03-15
 
 ### Added